feat(MaterialsAPI): use TypeScript
